backend/requirements.txt
FastAPI==0.115.12
uvicorn==0.34.2
httpx==0.28.1
aiohttp==3.11.18
pydantic==2.11.4
pydantic-ai==0.1.9
pydantic-settings==2.9.1
pydantic[email]
python-dotenv==1.1.0
python-multipart==0.0.20
python-jose==3.4.0
passlib==1.7.4
pymongo==4.12.1
jinja2==3.1.6
apscheduler==3.11.0

# --- testing extras

pytest==8.2.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
mongomock==4.1.2
PyJWT==2.9.0
respx==0.20.2  